Refactor `Collector` -> `Executor`.
