+++ mitiq/executor/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020 Unitary Fund
+# Copyright (C) 2021 Unitary Fund
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -13,6 +13,4 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from mitiq._typing import SUPPORTED_PROGRAM_TYPES, QPROGRAM
-from mitiq._about import about
-from mitiq._version import __version__
+from mitiq.executor.executor import Executor, generate_collected_executor
